backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.22

pytesseract==0.3.10
pypdf2==3.0.1
reportlab==4.0.7

pillow==10.1.0
pdf2image==1.16.3
python-dotenv==1.0.0
python-multipart==0.0.6
groq==0.9.0

